# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_case_media_api_cases__case_id__media_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "tracking_code": {
            "title": "Tracking Code",
            "type": "string"
          }
        },
        "required": [
          "tracking_code",
          "file"
        ],
        "title": "Body_upload_case_media_api_cases__case_id__media_post",
        "type": "object"
      },
...
      "CaseAssignmentRequest": {
        "properties": {
          "assignee_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "department": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Department",
            "type": "string"
          }
        },
        "required": [
          "department"
        ],
        "title": "CaseAssignmentRequest",
        "type": "object"
      },
      "CaseCreateRequest": {
        "properties": {
          "accuracy_m": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          },
          "category": {
            "enum": [
              "road",
              "waste",
              "water",
              "power",
              "public_building",
              "drainage",
              "safety",
              "other"
            ],
            "title": "Category",
            "type": "string"
          },
          "consent_version": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Consent Version",
            "type": "string"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 5,
            "title": "Description",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_confirmed": {
            "default": false,
            "title": "Location Confirmed",
            "type": "boolean"
          },
          "location_source": {
            "anyOf": [
              {
                "enum": [
                  "gps",
                  "manual_pin"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Source"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "priority": {
            "default": "normal",
            "enum": [
              "low",
              "normal",
              "high",
              "critical"
            ],
            "title": "Priority",
            "type": "string"
          },
          "privacy_consent": {
            "const": true,
            "title": "Privacy Consent",
            "type": "boolean"
          },
          "submitter_address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitter Address"
          },
          "submitter_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitter Name"
          },
          "submitter_phone": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitter Phone"
          },
          "village_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Village Id"
          }
        },
        "required": [
          "category",
          "description",
          "privacy_consent",
          "consent_version"
        ],
        "title": "CaseCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ba Na SmartLink API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cases": {
      "get": {
        "description": "Internal queue; Postgres RLS remains the final scope boundary.",
        "operationId": "list_cases_api_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Cases Api Cases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cases",
        "tags": [
          "citizen-cases"
        ]
      },
      "post": {
        "description": "Submit a public field report.  Citizens do not need an account.",
        "operationId": "create_case_api_cases_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Case Api Cases Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case",
        "tags": [
          "citizen-cases"
        ]
      }
    },
    "/api/cases/routing-rules": {
      "get": {
        "description": "Return the approved or demo routing catalogue within the caller's RLS scope.",
        "operationId": "list_routing_rules_api_cases_routing_rules_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Routing Rules Api Cases Routing Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Routing Rules",
        "tags": [
          "citizen-cases"
        ]
      }
    },
    "/api/cases/track/{tracking_code}": {
      "get": {
        "description": "Public, non-PII status lookup by an unguessable token.",
        "operationId": "track_case_api_cases_track__tracking_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "tracking_code",
            "required": true,
            "schema": {
              "title": "Tracking Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Track Case Api Cases Track  Tracking Code  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Case",
        "tags": [
          "citizen-cases"
        ]
      }
    },
    "/api/cases/{case_id}/assignment": {
      "post": {
        "description": "Admin-only routing mutation; AI suggestions can never call this endpoint.",
        "operationId": "assign_case_api_cases__case_id__assignment_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Assign Case Api Cases  Case Id  Assignment Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Case",
        "tags": [
          "citizen-cases"
        ]
      }
    },
    "/api/cases/{case_id}/media": {
      "post": {
        "description": "Store a citizen image only after proving possession of its tracking code.\n\nThe bucket is private and the service key is used only by this narrow,\nserver-validated path.  Public responses never expose the storage path.",
        "operationId": "upload_case_media_api_cases__case_id__media_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_case_media_api_cases__case_id__media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Case Media Api Cases  Case Id  Media Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Case Media",
        "tags": [
          "citizen-cases"
        ]
      }
    },
    "/api/cases/{case_id}/status": {
      "patch": {
        "operationId": "update_case_status_api_cases__case_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Case Status Api Cases  Case Id  Status Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Case Status",
        "tags": [
          "citizen-cases"
        ]
      }
    },
...
  }
}